refactor: promote logic loader upper
