+++ simaple/data/__init__.py
@@ -1,0 +1,6 @@
+from simaple.data._builtin import get_damage_logic
+
+
+__all__ = [
+    "get_damage_logic",
+]
