+++ simaple/data/jobs/readme.md
@@ -1,0 +1,5 @@
+### Jobs
+
+Declare job-level information here.
+
+May aggregated by jobs.
