+++ simaple/simulate/interface/simulator_configuration.py
@@ -5,7 +5,7 @@ import pydantic
 from simaple.container.simulation import SimulationContainer, SimulationSetting
 from simaple.core.base import ActionStat, Stat
 from simaple.core.jobtype import JobType
-from simaple.data.damage_logic import get_damage_logic
+from simaple.data import get_damage_logic
 from simaple.data.skill_profile import SkillProfile, get_skill_profile
 from simaple.simulate.base import SimulationRuntime
 from simaple.simulate.builder import EngineBuilder
